backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pytesseract==0.3.13
Pillow==10.4.0
PyPDF2==3.0.1
httpx==0.27.2
aiofiles==24.1.0
python-dotenv==1.0.1
